Feature/jwt route protection

+++ src/routes/users.py
@@ -2,6 +2,8 @@ from smtplib import SMTPException
 from flask import Blueprint, jsonify, request, current_app
 from util.database import Database
 from util.response import create_error_response
+from util.request import require_roles
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_jwt_extended import create_access_token
 from util.email import Emailer
 import re
@@ -228,11 +230,13 @@ def login(**kwargs):
 
 
 @users_blueprint.route("/<int:user_id>", methods=["DELETE"])
+@jwt_required()
 @Database.with_connection()
 def delete_user(user_id, **kwargs):
     cursor = kwargs["cursor"]
     connection = kwargs["connection"]
 
+    jwt_user = get_jwt_identity()
     incoming_data = request.get_json()
 
     if not incoming_data:
@@ -258,6 +262,11 @@ def delete_user(user_id, **kwargs):
         ):
             return create_error_response("Invalid credentials", 401)
 
+        if jwt_user["role"].lower() == "user" and jwt_user["ID"] != user_id:
+            return create_error_response(
+                "You don't have permission to view this resource", 403
+            )
+
         # sql query to delete user if it exists already
         query = "DELETE FROM users WHERE ID = %s"
 
@@ -273,6 +282,7 @@ def delete_user(user_id, **kwargs):
 
 
 @users_blueprint.route("/<int:user_id>", methods=["GET"])
+@require_roles(["admin", "super"])
 @Database.with_connection()
 def get_user_by_id(user_id, **kwargs):
     cursor = kwargs["cursor"]
@@ -299,6 +309,7 @@ def get_user_by_id(user_id, **kwargs):
 
 
 @users_blueprint.route("/", methods=["GET"])
+@require_roles(["admin", "super"])
 @Database.with_connection()
 def get_all_users(**kwargs):
     cursor = kwargs["cursor"]
@@ -323,6 +334,7 @@ def get_all_users(**kwargs):
 
 
 @users_blueprint.route("/<int:user_id>/role", methods=["PATCH"])
+@require_roles(["super"])
 @Database.with_connection()
 def update_user_role(user_id, **kwargs):
     cursor = kwargs["cursor"]
@@ -354,6 +366,7 @@ def update_user_role(user_id, **kwargs):
 
 
 @users_blueprint.route("/<int:user_id>/email", methods=["PATCH"])
+@jwt_required()
 @Database.with_connection()
 def send_update_email(user_id, **kwargs):
     """
@@ -361,10 +374,17 @@ def send_update_email(user_id, **kwargs):
     """
     cursor = kwargs["cursor"]
     connection = kwargs["connection"]
+
+    jwt_user = get_jwt_identity()
     request_data = request.get_json()
 
     if not request_data:
         return create_error_response("A body is required", 400)
+
+    if jwt_user["role"].lower() == "user" and jwt_user["ID"] != user_id:
+        return create_error_response(
+            "You don't have permission to view this resource", 403
+        )
 
     try:
         email = request_data["email"]
@@ -619,14 +639,21 @@ def update_user_email(**kwargs):
 
 
 @users_blueprint.route("<int:user_id>/updateName", methods=["PATCH"])
+@jwt_required()
 @Database.with_connection()
 def update_name(user_id, **kwargs):
     cursor = kwargs["cursor"]
     connection = kwargs["connection"]
     request_data = request.get_json()
+    jwt_user = get_jwt_identity()
 
     if not request_data:
         return create_error_response("A body is required", 400)
+
+    if jwt_user["role"].lower() == "user" and jwt_user["ID"] != user_id:
+        return create_error_response(
+            "You don't have permission to view this resource", 403
+        )
 
     try:
         full_name = request_data["fullName"]
